fix-migrationファイル

+++ db/migrate/20230915062433_add_star_to_recipes.rb
@@ -1,5 +1,5 @@
 class AddStarToRecipes < ActiveRecord::Migration[6.1]
   def change
-    add_column :recipes, :star, :float, null: false, default: 0
+    add_column :recipes, :star, :integer, null: false, default: 0
   end
 end
